Add feature for group submission
