+++ accounts/admin.py
@@ -4,7 +4,6 @@ from .models import *
 
 @admin.register(User)
 class UserAdmin(UserAdmin):
-    list_display= [f.name for f in User._meta.fields]
     list_filter = ['user_type']
     search_fields = ['id', 'studentprofile__index', 'studentprofile__assigned_class_code', 'studentprofile__name']
