+++ core/models.py
@@ -125,18 +125,3 @@ class Resource(models.Model):
     name = models.CharField(max_length=200)
     ## File Name Format: (Resource Section ID)_(Resource ID)_(Resource Name).(format)
     file = models.FileField(blank=True, null=True, default=None)
-
-class StudentGroup(models.Model):
-    # for student groups to work, you need the classroom code, the group number, 
-    # and the index number of the students
-    # uniquely identified by the classroom code and group number
-
-    class Meta:
-        constraints=[
-            models.UniqueConstraint(fields = ['classroom', 'group_number'], name = 'unique_identifier')
-        ]
-
-    classroom = models.ForeignKey(Classroom, on_delete=models.CASCADE)
-    group_number = models.PositiveSmallIntegerField()
-
-    member_indexes = models.JSONField(default=list)
