+++ student_core/admin.py
@@ -5,3 +5,8 @@ from .models import *
 @admin.register(Enroll)
 class Enroll(admin.ModelAdmin):
     search_fields = ['studentUserID', 'classroom', 'studentIndex', 'score']
+
+
+@admin.register(StudentGroup)
+class StudentGroup(admin.ModelAdmin):
+    search_fields = ['classroom']
